Remove static state from saved state overriding
