+++ androidshared/src/main/java/org/odk/collect/androidshared/system/SavedInstanceStateProvider.kt
@@ -1,0 +1,7 @@
+package org.odk.collect.androidshared.system
+
+import android.os.Bundle
+
+interface SavedInstanceStateProvider {
+    fun getState(savedInstanceState: Bundle?): Bundle?
+}
